Update allowed origins and minor code improvements
Enhanced Next.js config to dynamically resolve allowed origins for server actions based on environment variables and defaults. Disabled a TypeScript ESLint rule for next-env.d.ts, improved type safety in db.ts, fixed import in riverService tests, formatted test code, and casted COUNT result to int in cleanup-non-climate script.

+++ scripts/cleanup-non-climate.ts
@@ -24,7 +24,7 @@ export async function run(opts: { dryRun?: boolean; closePool?: boolean } = {}) 
       a.id, 
       a.title, 
       a.dek,
-      COUNT(ac.category_id) as category_count
+      COUNT(ac.category_id)::int as category_count
     FROM articles a
     INNER JOIN article_categories ac ON ac.article_id = a.id
     WHERE a.published_at >= now() - interval '30 days'
@@ -120,5 +120,6 @@ if (import.meta.url === `file://${process.argv[1]}`) {
     endPool().finally(() => process.exit(1))
   })
 }
+
